BREAKING: deprecate `std/http/http_errors`
